CRISTAL-465: Make Electron app work properly again
Seems like due to a recent Electron update, a timing issue now occurs when invoking an IPC request to load the configuration in the webview. The IPC listener is only set up **after** the page is ready, at which point it already tried to invoke the configuration loader.
This commit fixes the issue by moving the IPC listener setup earlier in the application's lifecycle, without waiting for the page to be ready.